fix: update IpInfo struct with u32 for ASN field
The ASN field in the IpInfo struct has been updated from u16 to u32 to accommodate larger values.

+++ crates/lib/api/src/liblaf/ip.rs
@@ -23,7 +23,7 @@ pub struct IpInfo {
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct Geo {
-    pub asn: u16,
+    pub asn: u32,
     pub country: String,
     pub country_code: String,
     pub country_flag: String,
